+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to the Ajrly Platform will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.6] - 2025-10-14
+
+**Author:** Nasr Aldin
+
+### fix
+
+fix: ssr runtime issue and build
+
+### File Changed
+
+- src/app.tsx
+- src/components/ServerHead.tsx
+- src/entry-server.tsx
+- src/utils/helpers/checks.ts
+- src/utils/helpers/converts.ts
+- src/utils/helpers/locale/userLocale.ts
+
+---
+
 ## [0.3.5] - 2025-10-14
 
 **Author:** Nasr Aldin